requirements.txt
polyinterface>=2.1.0
deprecation
python-kasa
